Ditch separate ModeltranslationPageMixin for now and keep functionality on BasePage
